Fixed initialzer's list
Fixed the constructor issue as no initializer's list was being provided. Also, after each attribute and method, a declaration of its visibility mode was written, so, now, it will check if the last mode was the same as the current mode and then according to that, will work around the data.